+++ signaling_logger.go
@@ -1,35 +1,9 @@
 package ayame
 
 import (
-	"fmt"
-	"os"
-	"time"
-
 	"github.com/rs/zerolog"
-	"gopkg.in/natefinch/lumberjack.v2"
 )
 
 func InitSignalingLogger(config *Config) (*zerolog.Logger, error) {
-	if f, err := os.Stat(config.LogDir); os.IsNotExist(err) || !f.IsDir() {
-		return nil, err
-	}
-
-	logPath := fmt.Sprintf("%s/%s", config.LogDir, config.SignalingLogName)
-
-	// https://github.com/rs/zerolog/issues/77
-	zerolog.TimestampFunc = func() time.Time {
-		return time.Now().UTC()
-	}
-
-	writer := &lumberjack.Logger{
-		Filename:   logPath,
-		MaxSize:    config.LogRotateMaxSize,
-		MaxBackups: config.LogRotateMaxBackups,
-		MaxAge:     config.LogRotateMaxAge,
-		Compress:   config.LogRotateCompress,
-	}
-
-	logger := zerolog.New(writer).With().Timestamp().Logger()
-
-	return &logger, nil
+	return InitLogger(config, config.SignalingLogName, "signaling")
 }